Type fixes; refactor common lib
