+++ packages/esm-patient-common-lib/src/orders/index.ts
@@ -1,2 +1,3 @@
-export * from './order-basket';
+export * from './useOrderBasket';
+export * from './postOrders';
 export * from './types';
