+++ packages/esm-patient-common-lib/src/orders/useOrderBasket.ts
@@ -1,30 +1,8 @@
-import { createGlobalStore, openmrsFetch, useStoreWithActions } from '@openmrs/esm-framework';
+import { useStoreWithActions } from '@openmrs/esm-framework';
 import type { OrderBasketItem } from './types';
 import { getPatientUuidFromUrl } from '../get-patient-uuid-from-url';
-import { OrderPost } from './types';
 import { useEffect } from 'react';
-
-// The order basket holds order information for each patient. The orders are grouped by `key`
-// so that different parts of the application may manage their own order lists. For example,
-// the medication order list might be grouped as `medications`, and might be managed by the
-// medication order basket panel.
-interface OrderBasketStore {
-  items: {
-    [patientUuid: string]: {
-      [grouping: string]: Array<OrderBasketItem>;
-    };
-  };
-  postDataPrepFunctions: {
-    [grouping: string]: PostDataPrepFunction;
-  }
-}
-
-export type PostDataPrepFunction = (order: OrderBasketItem, patientUuid: string, encounterUuid: string) => OrderPost;
-
-const orderBasketStore = createGlobalStore<OrderBasketStore>('order-basket', {
-  items: {},
-  postDataPrepFunctions: {},
-});
+import { OrderBasketStore, PostDataPrepFunction, orderBasketStore } from './store';
 
 const orderBasketStoreActions = {
   setOrderBasketItems(
@@ -122,38 +100,4 @@ export function useOrderBasket(grouping?: string | null, postDataPrepFunction?: 
     };
     return { orders, clearOrders, setOrders } as UseOrderBasketReturn<void>;
   }
-}
-
-export async function postOrders(
-  encounterUuid: string,
-  abortController: AbortController,
-) {
-  const patientUuid = getPatientUuidFromUrl();
-  const { items, postDataPrepFunctions }: OrderBasketStore = orderBasketStore.getState();
-  const patientItems = items[patientUuid];
-  
-  const erroredItems: Array<OrderBasketItem> = [];
-  for (let grouping in patientItems) {
-    const orders = patientItems[grouping];
-    for (let i = 0; i < orders.length; i++) {
-      const order = orders[i];
-      const dto = postDataPrepFunctions[grouping](order, patientUuid, encounterUuid);
-      await postOrder(dto, abortController).catch((error) => {
-        erroredItems.push({
-          ...order,
-          orderError: error,
-        });
-      });
-    }
-  }
-  return erroredItems;
-}
-
-function postOrder(body: OrderPost, abortController?: AbortController) {
-  return openmrsFetch(`/ws/rest/v1/order`, {
-    method: 'POST',
-    signal: abortController?.signal,
-    headers: { 'Content-Type': 'application/json' },
-    body,
-  });
 }
