+++ packages/esm-patient-orders-app/src/order-basket/order-basket-item-list.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Layer, Tile } from '@carbon/react';
 import { useTranslation } from 'react-i18next';
 import { useLayoutType } from '@openmrs/esm-framework';
-import type { OrderBasketItem } from '../types/order-basket-item';
+import type { OrderBasketItem } from '@openmrs/esm-patient-common-lib';
 import OrderBasketItemTile from './order-basket-item.component';
 import styles from './order-basket-item-list.scss';
