+++ packages/esm-patient-orders-app/src/order-basket/order-basket-item.component.tsx
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button, ClickableTile, Tile } from '@carbon/react';
 import { TrashCan, Warning } from '@carbon/react/icons';
-import type { OrderBasketItem } from '../types/order-basket-item';
+import type { OrderBasketItem } from '@openmrs/esm-patient-common-lib';
 import { useLayoutType } from '@openmrs/esm-framework';
 import styles from './order-basket-item.scss';
